(PC-25587) refactor(signup): Move step config to separate file
